rename symbolX to symbolTimes, alias as symbolX
