+++ src/symbol/times.js
