+++ test/symbols-test.js
@@ -1,5 +1,5 @@
 import assert from "assert";
-import {symbols, symbolsFill, symbolsStroke, symbolCircle, symbolCross, symbolDiamond, symbolSquare, symbolStar, symbolTriangle, symbolWye, symbolPlus, symbolX, symbolTriangle2, symbolAsterisk, symbolSquare2, symbolDiamond2} from "../src/index.js";
+import {symbols, symbolsFill, symbolsStroke, symbolCircle, symbolCross, symbolDiamond, symbolSquare, symbolStar, symbolTriangle, symbolWye, symbolPlus, symbolTimes, symbolTriangle2, symbolAsterisk, symbolSquare2, symbolDiamond2} from "../src/index.js";
 
 it("symbols is a deprecated alias for symbolsFill", () => {
   assert.strictEqual(symbols, symbolsFill);
@@ -21,7 +21,7 @@ it("symbolsStroke is the array of symbol types", () => {
   assert.deepStrictEqual(symbolsStroke, [
     symbolCircle,
     symbolPlus,
-    symbolX,
+    symbolTimes,
     symbolTriangle2,
     symbolAsterisk,
     symbolSquare2,
